requirements.txt
Flask==2.3.2
torch==2.0.1
torchvision==0.20.1
Pillow==10.0.1
ultralytics==8.3.178   # if it's a YOLOv8 model
